mcp-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "uac-ai-mcp"
version = "0.1.0"
description = "MCP (Model Context Protocol) server for UAC AI Parser forensic analysis platform"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "UAC AI Team" },
]
keywords = ["mcp", "forensics", "uac", "incident-response", "security"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "mcp[cli]>=1.2.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
]

[project.urls]
Repository = "https://github.com/your-org/uac-ai"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
uac-ai-mcp = "uac_ai_mcp.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["uac_ai_mcp"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
